Populate statement list for while node
There is a bug somewhere. As expected, the parser crashes in the body of
the while statement (since that assignment isn't parsed yet). What I
didn't expect is that the traceback shows that stmt_toplevel() is trying
to parse the body of the while... even though the call I'm making is to
stmt().