Lexer work -- making things more common
@jdpage, I saw your TODOs in the lexer and I agree. Not only does it
make the lexer simpler, the parser shouldn't need to have knowledge of
what the *expected* text of a node is, it only needs knowledge of the
*actual* text. And it was definitely a mistak to have hard-coded magic
strings in both the parser and the lexer. My bad on that.

I also removed the distinction between "Sprinkles" and "Keywords" in the
lexer. Now there is just "Known Words". The lexer shouldn't care whether
a node is a keyword or an operator, it just needs to know exactly which
node to emit.