Clean up binary munging, put exehdr in startup symbol
Instead of doing various binary-mangling directly, use the struct module from
the standard library.

Also, the startup symbol now directly includes the BASIC loading stub, which
simplifies the linker. This also allows us to parameterize the line number of
the basic code, to be used as a version number. In addition, the new calling
convention allows the startup code to be a byte shorter by JMPing directly to
main(), rather than JSRing to it, since main() will JSR anyway now.