Improve pattern-xform, implement ExplicitScopes, redo ResolveScopes
Patterns are now defined using a decorator on a generator. This is
mostly to make it easy to add docstrings. We now also have the ability
to match multiple child nodes with a single pattern object, the use of
which can be observed in the now-implemented ExplicitScopes.

Additionally, patterns can now be either ascending or descending.
(Before, only ascending was possible.) Ascending patterns attempt to
transform nodes as they exit them, while descending patterns attempt to
do so as they enter.