Replace ANTLR4 with homegrown LALR parser generator
This should be the last time I replace the parser.

Besides the change in the grammar, lexer has been rewritten (again) to
be stateless. This has been accomplished using "lexer modes", which are
a thing I saw mentioned in an Oil Shell blog post, and then came up with
my own implementation.

Also, the old "AstBuilder" class, which turned the ANTLR4 parse tree
into an AST, has been replaced with a "Simplify" pass, which turns the
output of the parser into the same AST format. One nice side-effect of
this is that, where if the AstBuilder didn't understand something, it
was left out of the tree entirely, the Simplify pass will pass it
through unchanged.

Of course, we lost the dependency on ANTLR4 -- we no longer require the
runtime, nor is there a separate build step after modifying the grammar.