Move AST to be directly under jeff65
When, during the ANTLR4 conversion, we switched to having a
fully-generic AST type instead of a smattering of types deeply tied into
the parser, this became generic enough that it's useful for multiple
frontends.

In addition, with the upcoming frontend/backend separation, it's
actually going to be needed by both submodules.