Store decendents of AstNodes exclusively in attrs (#57)
Ultimately this simplifies the code, especially sequence handling, since
it can now just be pattern matched as a linked list.

This was brought on by my reading of the Dragon Book, where I noticed
that it didn't distinguish between "children" and "attributes" of AST
nodes. This clicked with me, since I'd been struggling to determine what
should be a child vs. an attr since the AstNode structure was
introduced.

This also has a nice benefit that we no longer need the weird behavior
where if you return a list from transform_exit, it'll get spliced into
the child list of the parent. All transforms return exactly one node; if
you need to return multiple, match the containing list node and rewrite
the linked list appropriately.