Add a (Linux-only) test harness for gold-syntax programs.
Requires Xvfb and VICE to be installed. Additionally, it uses the pty
module, so it's Linux-only for now.

From a test-writing perspective, all that's required is adding a
decorator to the test, indicating a gold-syntax file. The decorator
arranges for the contents of the C64 memory after the file is run to be
passed to the test.

Functions by compiling the program, running it in VICE, and breaking at
program exit. This activates the VICE monitor, which allows us to send a
command to dump the contents of memory to a file, from whence we can
read it and pass the bytes to the test. The VICE monitor doesn't work in
console mode, so we run it under Xvfb to avoid the window popping up.