Replace hardcoded pass sequence with a scheduler
Passes are now tagged with information about what they do, and are
automatically placed into the correct order. The scheduling algorithm
is designed such that the same input will always produce the same
output, so that the order that passes tests is the same order used in
production. This can be disabled for testing purposes.